refactor pipes, idp table config
