+++ console/src/app/modules/idp-table/idp-table.component.scss
@@ -61,14 +61,22 @@ tr {
 .flex-row {
   display: flex;
   justify-content: space-between;
+  padding: 2px 0;
+  width: 250px;
 
   .key {
-    margin-right: 1rem;
-    font-size: 14px;
+    margin-right: .5rem;
+    font-size: 12px;
+    flex: 1 0 50%;
+    overflow: hidden;
+    text-overflow: ellipsis;
   }
 
   .value {
-    font-size: 14px;
+    font-size: 12px;
+    flex: 1 0 50%;
+    overflow: hidden;
+    text-overflow: ellipsis;
   }
 
   &:first-child {
