+++ console/src/app/pages/iam/iam-routing.module.ts
@@ -24,22 +24,27 @@ const routes: Routes = [
         },
     },
     {
-        path: 'idp/create',
-        loadChildren: () => import('src/app/modules/idp-create/idp-create.module').then(m => m.IdpCreateModule),
-        canActivate: [AuthGuard, RoleGuard],
-        data: {
-            roles: ['iam.idp.write'],
-            serviceType: PolicyComponentServiceType.ADMIN,
-        },
-    },
-    {
-        path: 'idp/:id',
-        loadChildren: () => import('src/app/modules/idp/idp.module').then(m => m.IdpModule),
-        canActivate: [AuthGuard, RoleGuard],
-        data: {
-            roles: ['iam.idp.read'],
-            serviceType: PolicyComponentServiceType.ADMIN,
-        },
+        path: 'idp',
+        children: [
+            {
+                path: 'create',
+                loadChildren: () => import('src/app/modules/idp-create/idp-create.module').then(m => m.IdpCreateModule),
+                canActivate: [AuthGuard, RoleGuard],
+                data: {
+                    roles: ['iam.idp.write'],
+                    serviceType: PolicyComponentServiceType.ADMIN,
+                },
+            },
+            {
+                path: ':id',
+                loadChildren: () => import('src/app/modules/idp/idp.module').then(m => m.IdpModule),
+                canActivate: [AuthGuard, RoleGuard],
+                data: {
+                    roles: ['iam.idp.read'],
+                    serviceType: PolicyComponentServiceType.ADMIN,
+                },
+            },
+        ],
     },
     {
         path: `policy/${PolicyComponentType.LOGIN}`,
