+++ console/src/app/pages/orgs/orgs-routing.module.ts
@@ -39,7 +39,6 @@ const routes: Routes = [
             },
         ],
     },
-
     {
         path: 'policy',
         children: [
