+++ console/src/app/pipes/has-role-pipe/has-role-pipe.module.ts
