+++ console/src/app/pipes/regexp-pipe/regexp.pipe.ts
